refactor: HttpRequest, HttpResponse 세분화 및 요청에 따라 Controller 실행
